requirements.txt
django
djangorestframework
markdown
django-filter
drf-access-policy
cerberus